# PluginResources/OpenApiPlugins/LightPlugin.swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Light API",
        "version": "v1"
    },
    "paths": {
        "/Light": {
            "get": {
                "tags": [
                    "Light"
                ],
                "summary": "Retrieves all lights in the system.",
                "operationId": "get_all_lights",
                "responses": {
                    "200": {
                        "description": "Returns a list of lights with their current state",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LightStateModel"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LightStateModel"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LightStateModel"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/Light/{id}": {
            "get": {
                "tags": [
                    "Light"
                ],
                "summary": "Retrieves a specific light by its ID.",
                "operationId": "get_light",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the light from the get_all_lights tool.",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the requested light",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/LightStateModel"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LightStateModel"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LightStateModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "If the light is not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Light"
                ],
                "summary": "Changes the state of a light.",
                "operationId": "change_light_state",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the light to change from the get_all_lights tool.",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The new state of the light and change parameters.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeStateRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeStateRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangeStateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns the updated light state",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/LightStateModel"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LightStateModel"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LightStateModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "If the light is not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ChangeStateRequest": {
                "type": "object",
                "properties": {
                    "isOn": {
                        "type": "boolean",
                        "description": "Specifies whether the light is turned on or off.",
                        "nullable": true
                    },
                    "hexColor": {
                        "type": "string",
                        "description": "The hex color code for the light.",
                        "nullable": true
                    },
                    "brightness": {
                        "type": "integer",
                        "description": "The brightness level of the light.",
                        "format": "int32",
                        "nullable": true
                    },
                    "fadeDurationInMilliseconds": {
                        "type": "integer",
                        "description": "Duration for the light to fade to the new state, in milliseconds.",
                        "format": "int32",
                        "nullable": true
                    },
                    "scheduledTime": {
                        "type": "string",
                        "description": "The time at which the change should occur.",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "additionalProperties": false,
                "description": "Represents a request to change the state of the light."
            },
            "LightStateModel": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "on": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "brightness": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "hexColor": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            }
        }
    }
}